pyproject.toml
[tool.poetry]
name = "gfr"
version = "0.1.0"
description = "Git Flow Assistant of Rahmasir"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["rich"], version = "^0.9.0"}
python-dotenv = "^1.0.0"
PyGithub = "^2.1.1"
questionary = "^2.0.1"
rich = "^13.7.0"

[tool.poetry.scripts]
ggg = "gfr.app:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
